Fix static resources loading

+++ src/main/java/pl/edu/agh/ki/speedgame/model/dao/Mark.java
@@ -21,7 +21,7 @@ public class Mark {
     private String name;
 
     @Column(nullable = false)
-    private int averageMark;
+    private double averageMark;
 
     @Column(nullable = false)
     private int amountOfMarks;
